Backend/todoenv/src/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-annoying==0.10.6
django-cors-headers==4.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
